requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
Django==5.0.1
django-ninja==1.1.0
django-types==0.19.1
packaging==23.2
psycopg==3.1.16
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
types-psycopg2==2.9.21.20
typing_extensions==4.9.0
tzdata==2023.4
